Simply plugin loading to bypass the Hairball constructor.
These changes require the offline.py file to directly call and load the plugins
which it ends up needing to do anyway.

Along with these changes are more-precise command-line arugment
parsing. Because the program `kelp` is provided upon installation (assuming
installed python programs are installed correctly on the system path) it no
longer matters where the plugins are located relative to the current working
directory. All that matters is that the plugins are installed along with the
package, hence using the `development` option.

Updated the README to reflect the changes in installation and usage.